Move suitableScope to common file
